refactor: 固定スタイルの切り分け
