+++ components/shared/Header/index.css.ts
@@ -1,4 +1,4 @@
-import { baseTransition, pcHeaderHeight, spHeaderHeight } from '@/consts'
+import { pcHeaderHeight, spHeaderHeight, baseTransition } from '@/consts/styles.css'
 import { style } from '@vanilla-extract/css'
 
 export const header = style({
